fix: UserDefaultsUtil 통합
