+++ Projects/Service/Sources/UserDefaultsUtil/UserDefaultsUtil.swift
@@ -7,6 +7,10 @@
 
 import Foundation
 
+public enum Key: String {
+    case recentKeyword
+}
+
 public struct UserDefaultsUtil {
     public static let shared = UserDefaultsUtil()
     
@@ -16,6 +20,15 @@ public struct UserDefaultsUtil {
     
     public enum UserDefaultKey: String {
         case userId
+        case recentKeyword
+    }
+    
+    public func setRecentKeywordList(_ dataList: [String]) {
+        defaults.setValue(dataList, forKey: UserDefaultKey.recentKeyword.rawValue)
+    }
+    
+    public func recentKeywordList() -> [String]? {
+        return defaults.value(forKey: UserDefaultKey.recentKeyword.rawValue) as? [String]
     }
     
     public func setUserId(_ id: Int) {
@@ -24,5 +37,9 @@ public struct UserDefaultsUtil {
     
     public func getInstallationId() -> Int {
         return defaults.value(forKey: UserDefaultKey.userId.rawValue) as? Int ?? 0
+    }
+    
+    public func remove(key: Key) {
+        UserDefaults.standard.removeObject(forKey: key.rawValue)
     }
 }
